building on sol playground and manually backing up, i hate my life

+++ programs/shagaJoe/src/instructions/end_rental.rs
@@ -1,8 +1,9 @@
 use anchor_lang::prelude::*;
 use crate::{RentalAccounts};
 use crate::states::{AffairState, Rental};
+use borsh::{BorshSerialize, BorshDeserialize};
 
-
+#[derive(BorshSerialize, BorshDeserialize)]
 pub enum RentalTerminationAuthority {
     Clockwork,
     Client,
@@ -14,11 +15,11 @@ pub fn handler(ctx: Context<RentalAccounts>, termination_by: RentalTerminationAu
     let escrow_account = &mut ctx.accounts.escrow;
     let rental_account = &mut ctx.accounts.rental;
 
-    // Step 1: Calculate the actual time server was used (in hours)
+    // Step 1: Calculate the actual time server was used (in hours) TODO: USE U64
     let clock = Clock::get()?;
     let current_time = clock.unix_timestamp;
     let actual_time = (current_time - rental_account.rental_start_time) / 3600;
-    let actual_payment = (actual_time * rental_account.rent_amount) as u64;
+    let actual_payment = actual_time * rental_account.rent_amount;
 
     // Step 3: Transfer the due payment to the lender (server)
     escrow_account.try_transfer_lamports(&ctx.accounts.lender, actual_payment)?;
